readITData/requirements.txt
python-dotenv==0.21.0
python-gitlab==3.11.0
